Changes after merge conflicts
