+++ src/server/routes/controllers/login.js
@@ -1,0 +1,78 @@
+const logger = require('../../logger')
+const UserModel = require('../../models/user')
+const Utilities = require('../../utilities')
+const statusCodes = require('./utilities/statusCodes')
+const msg = require('./utilities/responseMessages')
+
+const passport = require('passport')
+const LocalStrategy = require('passport-local').Strategy
+
+const LOGIN_STRATEGY = 'loginStratefy'
+
+function createPostLoginController() {
+	return (req, res, next) => {
+	
+		if(req.isAuthenticated()) {
+			return res.status(statusCodes.OK).send(msg.successes.LOGGED_USER)
+		}
+	
+		passport.authenticate(LOGIN_STRATEGY, function(err, user, info) {
+	
+			if (!user) { 
+				res.status(statusCodes.UNAUTHORIZED)
+				return res.send(Utilities.createError(msg.errors.UNAUTHORIZED)) }
+			
+			req.logIn(user, function(err) {
+				if (err) { 
+					logger.error(err) 
+					res.status(statusCodes.INTERNAL_SERVER)
+					return res.send(Utilities.createError(msg.errors.INTERNAL_SERVER))
+				}
+				return res.status(statusCodes.OK).send(msg.successes.LOGGED_USER)
+			})
+		})(req, res, next)
+	}
+}
+
+passport.serializeUser( (user, done) => {
+	done(null, user.id)
+} )
+
+passport.deserializeUser( (id, done) => {
+	UserModel.findById(id, (err, user) => {
+		done(err, user)
+	})
+})
+
+passport.use(LOGIN_STRATEGY, new LocalStrategy({
+	usernameField: 'username',
+	passwordField: 'password'
+}, 
+(username, password, done) => {
+
+	const index = {
+		username
+	}
+
+	UserModel.findOne(index).exec()
+		.then( (user) => {
+			const hash = Utilities.createSaltedHash(user.salt, password)
+
+			if (hash == user.password) {
+				return done(null, user)
+			}
+
+			return done(null, false, {message: msg.errors.UNAUTHORIZED_ACCESS})
+		})
+		.catch(err => {
+			logger.error(err)
+			return done(null, false, {message: msg.errors.INTERNAL_SERVER})
+		})
+
+	
+}))
+
+
+module.exports = {
+	post: createPostLoginController
+}
