+++ src/server/routes/controllers/utilities/responseMessages.js
@@ -1,10 +1,13 @@
 
 module.exports = {
 	errors: {
-		FAILED_TO_CREATE_USER: 'Failed to create the user'
+		FAILED_TO_CREATE_USER: 'Failed to create the user',
+		UNAUTHORIZED: 'Unauthorized access',
+		INTERNAL_SERVER: 'Error'
 	},
 	successes: {
-		USER_HAS_BEEN_CREATED: 'User has been created'
+		USER_HAS_BEEN_CREATED: 'User has been created',
+		LOGGED_USER: 'successfully authorization'
 	}
 	
 }
