--- src/server/routes/utilities/messages.js
@@ -1,11 +1,0 @@
-
-module.exports = {
-	Errors: { 
-		BAD_REQUST: 'Bad request',
-		UNAUTHORIZED_ACCESS: 'Unauthorized access',
-		INTERNAL_SERVER: 'Internal server error'
-	},
-	Success: {
-		LOGGED_USER: 'You are logged now'
-	}
-}
